UPDATED AdminPage.php
Added SQL functions (borrowed from Gareth's UserProfile.php page). Also added some comments of what we need to do to get the page to work as intended. Username and password for the person logged in are also pulled (this will be changed to pull the details from the user selected in the drop down). Sorry this isnt a one line changelog but i did a few things and needed to explain them properly lol.