UPDATED MainActivity.txt
Changed the button text. When the button is pressed the data: Battery Level' and 'DeviceID' (Phone's name, got from the 'Build.MODEL' code is sent to the phpmyadmin database.
At this current point in time the code just sends the charge state as '1' as the code for this isnt working just yet.