UPDATED activity_main.txt
Added text to tell the user what their device ID is. I don't know if they need to know this when setting up there account. Can remove if need be. Also changed the default text for each text view, although this text is only visible inside the design for the app not when it comes to run time.